requirements.txt
boto==2.40.0
Django==1.9.6
django-storages==1.4.1
